Moved tests to implementation.
